# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ops-service API",
        "title": "ops-service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/ops-service",
    "paths": {
...
        "/monitor/gateway/detail": {
            "get": {
                "description": "监控接口-gateway-detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控接口-gateway-detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "identifier",
                        "name": "identifier",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.MonitorGatewayDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/monitor/list/dbs": {
            "get": {
                "description": "监控数据库列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控数据库列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "过滤名称, 不填则为全部",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.MonitorListDbItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/monitor/list/gateways": {
            "get": {
                "description": "监控采集网关列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控采集网关列表",
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.MonitorListGatewayItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/monitor/list/hosts": {
            "get": {
                "description": "监控主机列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控主机列表",
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.MonitorListHostItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/monitor/list/interfaces": {
            "get": {
                "description": "监控接口列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控接口列表",
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.MonitorListInterfaceItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/monitor/list/services": {
            "get": {
                "description": "监控服务列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控服务列表",
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.MonitorListServiceItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/monitor/log": {
            "get": {
                "description": "监控接口-log",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控接口-log",
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.MonitorLog"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/monitor/log/detail": {
            "get": {
                "description": "监控接口-log-detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控接口-log-detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "级别， debug,info,error,warn,不填则为全部",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "过滤,可使用正则表达式",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制条数，默认100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "多少小时前，默认24",
                        "name": "pre_hours",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.MonitorLog"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/monitor/log/label/values": {
            "get": {
                "description": "监控接口-label values",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控接口-log-label values",
                "parameters": [
                    {
                        "type": "string",
                        "description": "label",
                        "name": "label",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/monitor/total": {
            "get": {
                "description": "监控总览",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "监控总览",
                "responses": {
                    "200": {
                        "description": "成功code和成功信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.MonitorTotal"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "entity.MonitorGatewayDetail": {
            "type": "object",
            "properties": {
                "identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.MonitorGatewayDetailPoint"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "entity.MonitorGatewayDetailPoint": {
            "type": "object",
            "properties": {
                "access_type": {
                    "description": "接入类型 ， IO/485",
                    "type": "string"
                },
                "address_485": {
                    "description": "如果是485，则是地址",
                    "type": "string"
                },
                "address_io": {
                    "description": "如果是io，则是AI1,DI1",
                    "type": "string"
                },
                "device_identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "状态，1：正常，0：异常",
                    "type": "integer"
                },
                "update_time": {
                    "description": "格式 2006-01-02 15:04:05",
                    "type": "string"
                },
                "value": {
                    "description": "值",
                    "type": "string"
                }
            }
        },
        "entity.MonitorListDbItem": {
            "type": "object",
            "properties": {
                "cache_hit_rate": {
                    "description": "缓存命中率",
                    "type": "number"
                },
                "conflict_count": {
                    "description": "冲突数",
                    "type": "integer"
                },
                "connection_count": {
                    "description": "连接数",
                    "type": "integer"
                },
                "deadlock_count": {
                    "description": "死锁数",
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "server": {
                    "description": "服务器",
                    "type": "string"
                },
                "status": {
                    "description": "运行状态，1：正常，0：停止",
                    "type": "integer"
                }
            }
        },
        "entity.MonitorListGatewayItem": {
            "type": "object",
            "properties": {
                "ai_port_count": {
                    "description": "ai端口数",
                    "type": "integer"
                },
                "ao_port_count": {
                    "description": "ao端口数",
                    "type": "integer"
                },
                "di_port_count": {
                    "description": "di端口数",
                    "type": "integer"
                },
                "do_port_count": {
                    "description": "do端口数",
                    "type": "integer"
                },
                "identifier": {
                    "description": "标识符",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "status": {
                    "description": "运行状态，1：正常，0：停止",
                    "type": "integer"
                }
            }
        },
        "entity.MonitorListHostItem": {
            "type": "object",
            "properties": {
                "cpu_usage": {
                    "description": "cpu使用率",
                    "type": "number"
                },
                "ip": {
                    "description": "ip",
                    "type": "string"
                },
                "kernel": {
                    "type": "string"
                },
                "mem_usage": {
                    "description": "内存使用率",
                    "type": "number"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "status": {
                    "description": "状态,1:正常,0:离线",
                    "type": "integer"
                }
            }
        },
        "entity.MonitorListInterfaceItem": {
            "type": "object",
            "properties": {
                "alert_count": {
                    "description": "告警次数",
                    "type": "integer"
                },
                "call_frequency": {
                    "description": "调用频率",
                    "type": "integer"
                },
                "latest_data_time": {
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "protocol": {
                    "description": "接口协议",
                    "type": "string"
                },
                "status": {
                    "description": "状态，1：正常，0：异常",
                    "type": "integer"
                },
                "third_party_system": {
                    "description": "第三方系统",
                    "type": "string"
                }
            }
        },
        "entity.MonitorListServiceItem": {
            "type": "object",
            "properties": {
                "host": {
                    "description": "主机",
                    "type": "string"
                },
                "memory_usage": {
                    "description": "占用内存数",
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "status": {
                    "description": "运行状态，1：正常，0：停止",
                    "type": "integer"
                },
                "thread_count": {
                    "description": "线程数",
                    "type": "integer"
                }
            }
        },
        "entity.MonitorLog": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                }
            }
        },
        "entity.MonitorTotal": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.MonitorTotalItem"
                    }
                }
            }
        },
        "entity.MonitorTotalItem": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "异常数",
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "normal": {
                    "description": "正常数",
                    "type": "integer"
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
    }
}